+++ Source/SourceKittenFramework/Module.swift
@@ -72,7 +72,13 @@ public struct Module {
     - parameter path:                Path to run `xcodebuild` from. Uses current path by default.
     */
     public init?(xcodeBuildArguments: [String], name: String? = nil, inPath path: String = FileManager.default.currentDirectoryPath) {
-        let name = name ?? moduleName(fromArguments: xcodeBuildArguments)
+        let buildSettings = XcodeBuild.showBuildSettings(arguments: xcodeBuildArguments, inPath: path)
+
+        let name = name
+            // Check for user-defined "SWIFT_MODULE_NAME", otherwise use "PRODUCT_MODULE_NAME".
+            ?? buildSettings?.firstBuildSettingValue { $0.swiftModuleName ?? $0.productModuleName }
+            ?? moduleName(fromArguments: xcodeBuildArguments)
+
         // Executing normal build
         fputs("Running xcodebuild\n", stderr)
         if let output = XcodeBuild.run(arguments: xcodeBuildArguments, inPath: path),
@@ -83,8 +89,7 @@ public struct Module {
         }
         // Check New Build System is used
         fputs("Checking xcodebuild -showBuildSettings\n", stderr)
-        if let output = XcodeBuild.run(arguments: xcodeBuildArguments + ["-showBuildSettings"], inPath: path),
-            let projectTempRoot = parseProjectTempRoot(xcodebuildOutput: output),
+        if let projectTempRoot = buildSettings?.firstBuildSettingValue(for: { $0.projectTempRoot }),
             let arguments = checkNewBuildSystem(in: projectTempRoot, moduleName: name),
             let moduleName = moduleName(fromArguments: arguments) {
             self.init(name: moduleName, compilerArguments: arguments)
@@ -130,5 +135,22 @@ extension Module: CustomStringConvertible {
     /// A textual representation of `Module`.
     public var description: String {
         return "Module(name: \(name), compilerArguments: \(compilerArguments), sourceFiles: \(sourceFiles))"
+    }
+}
+
+// MARK: XcodeBuildSetting Conveniences
+
+private extension Collection where Element == XcodeBuildSetting {
+    /// Iterates through the `XcodeBuildSetting`s and returns the first value returned by the getter closure.
+    ///
+    /// For example, if we want the value of the first `XcodeBuildSetting` with a `"PROJECT_TEMP_ROOT"` value:
+    ///
+    ///     let buildSettings: [XcodeBuildSetting] = ...
+    ///     let projectTempRoot = buildSettings.firstBuildSettingValue { $0.projectTempRoot }
+    ///
+    /// - Parameter getterClosure: A closure that returns a dynamic member.
+    /// - Returns: The first value returned by the getter closure.
+    func firstBuildSettingValue(for getterClosure: (XcodeBuildSetting) -> String?) -> String? {
+        return lazy.compactMap(getterClosure).first
     }
 }